Terms&Conditions and LegalNotice views ready
