+++ KarhooUISDK/UIComponents/TermsConditionsView/SwiftUI/KarhooTermsConditionsView.swift
@@ -33,87 +33,8 @@ struct KarhooTermsConditionsView: View {
             }
 
             TextView($viewModel.attributedText)
+                .isEditable(false)
                 .accessibilityValue(viewModel.accessibilityText)
         }
-    }
-}
-
-class KarhooTermsConditionsViewModel: ObservableObject {
-
-    // MARK: - Dependencies
-
-    private let sdkConfiguration: KarhooUISDKConfiguration
-
-    // MARK: - Properties
-
-    private var cancellables: Set<AnyCancellable> = []
-    var isAcceptanceRequired: Bool { sdkConfiguration.isExplicitTermsAndConditionsConsentRequired }
-
-    @Published var attributedText: NSAttributedString = .init(string: "")
-    @Published var accessibilityText: String = ""
-    @Published var confirmed: Bool = false { didSet {
-        updateImageName()
-    }}
-    @Published var imageName: String = "kh_uisdk_checkbox_selected"
-
-    // MARK: - Lifecycle
-
-    init(
-        sdkConfiguration: KarhooUISDKConfiguration = KarhooUISDKConfigurationProvider.configuration,
-        supplier: String?,
-        termsStringURL: String
-    ) {
-        self.sdkConfiguration = sdkConfiguration
-        self.setBookingTerms(supplier: supplier, termsStringURL: termsStringURL)
-        self.setupBinding()
-    }
-
-    // MARK: - Setup
-
-    private func setBookingTerms(supplier: String?, termsStringURL: String) {
-        let text = TermsConditionsStringBuilder()
-            .bookingTermsCopy(
-                supplierName: supplier,
-                termsURL: convert(termsStringURL)
-            )
-        setText(text)
-    }
-
-    private func setupBinding() {
-        $confirmed
-            .sink { [weak self] _ in
-                self?.updateImageName()
-            }
-            .store(in: &cancellables)
-    }
-
-    // MARK: - Helpers
-
-    private func convert(_ stringURL: String) -> URL {
-        URL(string: stringURL) ?? TermsConditionsStringBuilder.karhooTermsURL()
-    }
-
-    private func setText(_ attributedText: NSAttributedString) {
-        let text = NSMutableAttributedString(attributedString: attributedText).then {
-            let paragraphStyle = NSMutableParagraphStyle()
-            paragraphStyle.alignment = isAcceptanceRequired ? .left : .center
-            $0.addAttribute(
-                .paragraphStyle,
-                value: paragraphStyle,
-                range: NSRange.init(location: 0, length: $0.length)
-            )
-        }
-        self.attributedText = text
-        self.accessibilityText = text.string.replacingOccurrences(of: "|", with: ".")
-    }
-
-    private func updateImageName() {
-        var newImageName: String {
-            switch confirmed {
-            case true: return "kh_uisdk_checkbox_selected"
-            case false: return "kh_uisdk_checkbox_unselected"
-            }
-        }
-        imageName = newImageName
     }
 }
